Revert the FLAC speedups; keep only the measurement
The optimisation round made things worse: every file now hangs, including the
one that previously played clean. So the D0/O14 readings taken on it are not
evidence about the decoder -- they describe a build that does not work.

Rather than fix forward through three suspicious changes at once, this restores
fw/flac.c, fw/flac.h and fw/build.sh byte-identical to 3f7f361, the build known
to play Pink Floyd cleanly, and keeps ONLY the instrumentation. The speedups
are preserved in 6fced2c and can come back one at a time once there is a
trustworthy baseline to compare against.

Of the three, two had no business shipping together while chasing a regression:

  - __attribute__((optimize("O2"))), which GCC's own documentation calls
    unsuitable for production code. It nearly doubled flac.o, 6958 -> 11926
    bytes, and forced always_inline expansion of to16() into a subframe_stream
    that instantiates the EMIT macro eight times.
  - a uint32_t copy out of the ring, which is a strict-aliasing violation
    against a uint8_t buffer at -O2, on a DMA target the rest of the file
    reaches through a VOLATILE pointer.

What is added instead is the ability to tell the causes apart:

  - E, the output path (fade, FIFO poll, R_AUDIO write) excluding the wait, so
    decode can be separated from the per-sample MMIO cost: decode = 100-D-O-E.
  - The MP3 FIFO wait now feeds the SAME idle counter. MP3 at 1x needs ~45.7 of
    60 MHz, so D should read about 20 on an MP3 track. That makes MP3 a CONTROL:
    if D reads 0 there too, the instrument is wrong and not the decoder -- which
    is exactly the doubt that made the last four readings unusable.

+++ fw/flac.h
@@ -51,7 +51,7 @@ typedef struct {
     void         *ctx;
     uint8_t       buf[512];
     uint32_t      have, pos;      /* bytes in buf, and read cursor          */
-    uint64_t      bitacc;         /* bit reservoir, refilled 4 bytes at a time */
+    uint32_t      bitacc;         /* bit reservoir                          */
     uint32_t      bitcnt;
     int           eof;
